Adds sassy styles to navbar
I set up scss files for navbar, colors, icons, as well as creating
an scss file that over-writes some of bootstrap's default navbar
responsiveness. I didn't want, for example, a hamburger (with dropdown)
to appear on small screens. Instead, I have icons for user profile
and user shopping cart, which remain on screen at all times. Next, I
want to attach side-drawer functionality to the shopping cart icon,
and a dropdown with 'sign-out' and 'change-password' functionality
to the user profile icon. Doing so should complete the basic
scaffold we need to build out the applications functionality.