Adds side-drawer to act as shopping cart
Adds functionality, so that when user clicks on the shopping
cart, a side-drawer pops out, showing the products a user
has added to the cart. From the cart, a user should be able to
see the product name, the price, and they should be able to
specify a quantity. The user should also be able to remove
the item from their shopping list, and submit the shopping cart
as an order.