Creates cart object and renders that object
The cart is now rendered using an object stored in
store.cart.products, where products is an array containing
many product objects.  I also improved the css for how products
render in the cart.  Next, I need to take the cart object and
reconfigure it so that it can be sent as data to our orders
collection/route.