Removes from cart / displays message upon adding
When a user adds a fruit to the cart, the cart now opens,
and a message displays for 1 second at the top fo the page
telling the user the item has been added.

I also added the ability for a user to remove
a piece of fruit from the cart. Upon clicking the x
next to an item in the cart, that fruit is removed
from the cart object in store, and the div is visually
removed from the cart.